requirements.txt
Cerberus==1.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
flask-heroku==0.1.9
flask-marshmallow==0.10.0
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.2
psycopg2-binary==2.8.2
PyJWT==1.7.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.3
Werkzeug==0.15.2
